fix: split up scrapers into different modules
