+++ adapters/outgoing/scrapers/pom.xml
@@ -10,6 +10,7 @@
         <relativePath>../../pom.xml</relativePath>
     </parent>
     <artifactId>scrapers</artifactId>
+    <packaging>pom</packaging>
     <dependencies>
         <dependency>
             <groupId>org.jsoup</groupId>
