--- adapters/outgoing/scrapers/src/test/java/be/xplore/pricescraper/scrapers/LidlBeScraperTests.java
@@ -1,34 +1,0 @@
-package be.xplore.pricescraper.scrapers;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import be.xplore.pricescraper.item.LidlBeScraper;
-import be.xplore.pricescraper.item.Scraper;
-import java.io.IOException;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = LidlBeScraper.class)
-public class LidlBeScraperTests {
-  private final Scraper scraper;
-
-  public LidlBeScraperTests(@Qualifier("scraper-lidl.be") Scraper scraper) {
-    this.scraper = scraper;
-  }
-
-  @Test
-  void constructed() {
-    assertNotNull(scraper);
-  }
-
-  @Test
-  void getItemResults() throws IOException {
-    var response = scraper.scrape("bleekwater/p740192819");
-    assertTrue(response.isPresent());
-    assertNotNull(response.get());
-    assertNotNull(response.get().title());
-    assertTrue(response.get().price() > 0);
-  }
-}
