+++ adapters/pom.xml
@@ -16,8 +16,13 @@
         <module>outgoing</module>
         <module>incoming/rest</module>
         <module>outgoing/jpa</module>
-        <module>outgoing/scrapers</module>
         <module>matchers</module>
+        <module>outgoing/scrapers/ahscraper</module>
+        <module>outgoing/scrapers/scraper</module>
+        <module>outgoing/scrapers/carrefourscraper</module>
+        <module>outgoing/scrapers/aldiscraper</module>
+        <module>outgoing/scrapers/lildscraper</module>
+        <module>outgoing/scrapers/localdummyscraper</module>
     </modules>
 
     <properties>
